Optimized video loading for mp4 seek and play (#23)
This adds support for extracting mp4 segment information before loading video into Tator for optimized seek and play capability.  This allows for more streamlined interaction when previewing and editing results.  This requires the addition of the mp4dump tool from https://www.bento4.com/.  Additional simplifications were made in this PR, including removing explicit configuration of the ffmpeg path in the config.yml and consistent config.yml for all video loads in the test suite.   IMPORTANT: segments are stored in JSON format in same directory as the video, so that directory must be writeable by the user running the load.